Create Union.c
What is Union ? 
A union in C is a special data type that allows storing different data types in the same memory location. It is similar to a struct, but with a key difference: all members of a union share the same memory space, meaning only one member can hold a value at a time.